report/scripts/requirements.txt
docxcompose
pandas
python-docx
openpyxl
requests
